Complete replacement with openstacksdk
